requirements.txt
pygame==2.6.1
PyOpenGL==3.1.7
numpy==1.26.0